+++ products/core.test.browser.js
@@ -2956,6 +2956,14 @@ testTree.fromShape = equal => {
  1`
   )
 }
+testTree.getFrom = equal => {
+  // Arrange
+  const treeNode = new TreeNode(`name
+ string title The book of
+ string person Jai`)
+  // Act/Assert
+  equal(treeNode.nodeAt(0).getFrom("string person"), "Jai")
+}
 testTree.toOutline = equal => {
   // Arrange
   const treeNode = new TreeNode(`hello